refactor: extract MCP resource handlers into main_resources.go
